Camera board and main board built and uploaded
Resolving esp-now issues but good state to save
