+++ platformio.ini
@@ -41,6 +41,10 @@ build_src_filter =
     -<*>
     +<camera_board_main.cpp>
     +<utils/WiFiUtil.h>
+    +<utils/WiFiUtil.cpp>
+    +<config/Credentials.h>
+    +<config/Credentials.cpp>
+    +<utils/MessageTypes.h>
 monitor_speed = 115200
 build_unflags = -std=gnu++11
 build_flags = -std=gnu++17
