+++ src/config/Credentials.h.template
@@ -11,18 +11,11 @@
 #define VIDEO_WEB_SERVER_IP "YOUR_LAPTOP_IP"
 #define VIDEO_WEB_SERVER_PORT 5005
 
-#define MAIN_BOARD_IP "YOUR_MAIN_BOARD_IP"
-#define CAMERA_BOARD_IP "YOUR_CAMERA_BOARD_IP"
-
 #define FIREBASE_WEB_API_KEY "YOUR_FIREBASE_WEB_API_KEY"
 #define FIREBASE_USER_EMAIL "YOUR_FIREBASE_EMAIL"
 #define FIREBASE_USER_PASSWORD "YOUR_FIREBASE_PASSWORD"
 #define FIREBASE_DATABASE_URL "YOUR_FIREBASE_DATABASE_URL"
 
-// Board MAC addresses (replace with your actual MAC addresses)
+// Board MAC addresses (replace with your actual MAC addresses) but move to cpp file
 const uint8_t MAIN_BOARD_MAC_ADDRESS[6] = {0x00, 0x00, 0x00, 0x00, 0x00, 0x00};
 const uint8_t CAMERA_BOARD_MAC_ADDRESS[6] = {0x00, 0x00, 0x00, 0x00, 0x00, 0x00};
-
-#if MAIN_BOARD_MAC_ADDRESS[0] == 0x00 || CAMERA_BOARD_MAC_ADDRESS[0] == 0x00
-#error "Please copy Credentials.h.template to Credentials.h and set your MAC addresses"
-#endif
