+++ src/utils/MessageTypes.h
@@ -1,0 +1,6 @@
+#pragma once
+
+struct struct_message {
+  char message[32];
+  int camera_action;
+};
